Refactor employee role into own module
