+++ src/employee/employee.module.ts
@@ -3,7 +3,7 @@ import { EmployeeService } from './employee.service';
 import { EmployeeController } from './employee.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Employee } from './entities/employee.entity';
-import { EmployeeRole } from './entities/employee-role.entity';
+import { EmployeeRole } from '../employee-role/entities/employee-role.entity';
 
 @Module({
   imports: [
